Move run-script action to separate folder
