+++ script/__tests__/index.test.ts
@@ -1,0 +1,28 @@
+/**
+ * Copyright 2020-2023 Zowe Actions Contributors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import * as fs from "fs";
+import { loadScript } from "../src/loader";
+
+describe("Run Script action", () => {
+    const scriptNames = fs.readdirSync(__dirname + "/../scripts").map(s => s.slice(0, s.lastIndexOf(".")));
+
+    for (const scriptName of scriptNames) {
+        it("should load script " + scriptName, () => {
+            expect(typeof loadScript(scriptName)).toBe("function");
+        });
+    }
+});
