+++ script/scripts/prepareRelease.ts
