+++ script/scripts/sonarConfig.ts
