+++ script/src/index.ts
