+++ script/src/loader.ts
@@ -14,15 +14,20 @@
  * limitations under the License.
  */
 
-import * as fs from "fs";
-import { loadScript } from "../src/loader";
+import { IContext } from "@octorelease/core";
 
-describe("Run Script action", () => {
-    const scriptNames = fs.readdirSync(__dirname + "/../../../script/src").map(s => s.slice(0, s.lastIndexOf(".")));
+const SCRIPTS: { [key: string]: any } = {
+    npmUpdate: require("../scripts/npmUpdate"),
+    prepareRelease: require("../scripts/prepareRelease"),
+    sonarConfig: require("../scripts/sonarConfig")
+};
 
-    for (const scriptName of scriptNames) {
-        it("should load script " + scriptName, () => {
-            expect(typeof loadScript(scriptName)).toBe("function");
-        });
+// List of scripts that should only run in release branches
+export const RELEASE_SCRIPTS: string[] = ["npmUpdate"];
+
+export function loadScript(scriptName: string): (context: IContext) => Promise<void> {
+    if (!Object.keys(SCRIPTS).includes(scriptName)) {
+        throw new Error(`Could not find script to run: ${scriptName}`);
     }
-});
+    return SCRIPTS[scriptName].default;
+}
